requirements.txt
seaborn==0.12.2
opencv-python~=4.8.0.76
flask~=2.2.5
torch~=2.0.1+cu117
torchvision~=0.15.2+cu117
pillow~=10.0.0
numpy~=1.25.0